# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rede Akiba API REST",
        "description": "Esta documentação descreve as todas as operações disponíveis para as aplicações da Rede Akiba.",
        "contact": {
            "email": "contato@redeakiba.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/podcasts": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Retorna uma lista de todos os podcasts cadastrados",
                "description": "Retorna uma lista de todos os podcasts cadastrados no sistema da Rede Akiba.",
                "operationId": "d5212a7244ded3a2774904c0dd18a715",
                "responses": {
                    "200": {
                        "description": "Lista de todos os podcasts cadastrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum podcast encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhum podcast encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Cadastra um novo podcast",
                "description": "Cadastra um novo podcast no sistema da Rede Akiba",
                "operationId": "305313a82310209b76f4aeb5a0a5ca1b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PodcastRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podcast cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Autor não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Autor não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/podcasts/{slug}": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Retorna um podcast específico",
                "description": "Retorna um podcast específico cadastrado no sistema da Rede Akiba.",
                "operationId": "18f5ef0fb089c5d2fd24c0833172d813",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug do Podcast: Retorna um podcast baseado no slug fornecido",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podcast encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Podcast não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Podcast não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/podcasts/{id}": {
            "delete": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Remove um podcast específico",
                "description": "Remove um podcast específico cadastrado no sistema da Rede Akiba.",
                "operationId": "1408d1fd3c7cef2c7b6318dfb312e11b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Podcast: Remove um podcast baseado no ID fornecido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podcast removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Podcast removido com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Podcast não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Podcast não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Atualiza um podcast específico",
                "description": "Atualiza um podcast específico cadastrado no sistema da Rede Akiba.",
                "operationId": "260a08f682d5ff817f9e4381865bdc5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Podcast: Atualiza um podcast baseado no ID fornecido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PodcastRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podcast atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Podcast ou autor não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Podcast ou autor não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PodcastResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "season": {
                        "type": "integer"
                    },
                    "episode": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "resume": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "player": {
                        "type": "string"
                    },
                    "aggregators": {
                        "properties": {
                            "spotify": {
                                "type": "string"
                            },
                            "apple_podcasts": {
                                "type": "string"
                            },
                            "google_podcasts": {
                                "type": "string"
                            },
                            "anchor": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PodcastRequest": {
                "properties": {
                    "author": {
                        "type": "integer"
                    },
                    "season": {
                        "type": "integer"
                    },
                    "episode": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "resume": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "player": {
                        "type": "string"
                    },
                    "aggregators": {
                        "properties": {
                            "spotify": {
                                "type": "string"
                            },
                            "apple_podcasts": {
                                "type": "string"
                            },
                            "google_podcasts": {
                                "type": "string"
                            },
                            "anchor": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "access_levels": {
                        "type": "json",
                        "example": "['string', 'string']"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "age": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "biography": {
                        "type": "string"
                    },
                    "social_networks": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "facebook": {
                                    "type": "string",
                                    "example": "http://facebook.com/username"
                                },
                                "instagram": {
                                    "type": "string",
                                    "example": "http://instagram.com/username"
                                },
                                "x": {
                                    "type": "string",
                                    "example": "http://x.com/username"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "likes": {
                        "properties": {
                            "likes": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": "['string', 'string']"
                            },
                            "dislikes": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": "['string', 'string']"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}